reuse react-aria scroll gutter solution in select components, fix focus styling bug
