+++ src/components/pagination/pagination-popover.tsx
@@ -19,6 +19,7 @@ import {
 } from "react-aria";
 import { OverlayTriggerState, useOverlayTriggerState } from "react-stately";
 import { DOMProps } from "@react-types/shared";
+import { useReactAriaScrollGutterHack } from "src/hooks/useReactAriaScrollGutterHack";
 
 function PopupContents(
 	props: Pick<PaginationProps, "page" | "getPageHref" | "softNavigate"> & {
@@ -139,42 +140,13 @@ function PaginationPopover({
 	const { dialogProps, titleProps } = useDialog(overlayProps, dialogRef);
 	const { isFocusVisible } = useFocusVisible();
 
-	/**
-	 * bandaid solution for layout shift
-	 *
-	 * https://github.com/adobe/react-spectrum/issues/5470
-	 * https://github.com/adobe/react-spectrum/issues/1216
-	 * TODO: remove this padding-right whenever we have a better solution or react aria fixes the issue
-	 */
-	useLayoutEffect(() => {
-		const updateStyles = () => {
-			if (CSS.supports("scrollbar-gutter: stable")) {
-				document.documentElement.style.paddingRight = "0";
-			}
-		};
-
-		// immediately invoke to set the styles we want
-		updateStyles();
-
-		// Observe attribute changes to apply styles as needed
-		const mutationObserver = new MutationObserver(updateStyles);
-		mutationObserver.observe(document.documentElement, {
-			attributes: true,
-			attributeFilter: ["style"],
-			subtree: false,
-			attributeOldValue: false,
-		});
-
-		return () => {
-			// Clean up observer on component unmount
-			mutationObserver.disconnect();
-		};
-	}, []);
+	// bandaid solution for layout shift
+	useReactAriaScrollGutterHack();
 
 	return (
 		<Overlay>
 			<div {...underlayProps} className={style.underlay} />
-			
+
 			<div {...popoverProps} ref={popoverRef} className={style.popup}>
 				<svg
 					width="24"
